+++ ext/numo/narray/src/t_dcomplex.c
@@ -1,4 +1,3 @@
-
 /*
   t_dcomplex.c
   Ruby/Numo::NArray - Numerical Array class for Ruby
@@ -4944,7 +4943,7 @@ static VALUE dcomplex_poly(VALUE self, VALUE args) {
   ndfunc_arg_out_t aout[1] = {{cT, 0}};
   ndfunc_t ndf = {iter_dcomplex_poly, NO_LOOP, 0, 1, 0, aout};
 
-  argc = RARRAY_LEN(args);
+  argc = (int)RARRAY_LEN(args);
   ndf.nin = argc + 1;
   ndf.ain = ALLOCA_N(ndfunc_arg_in_t, argc + 1);
   for (i = 0; i < argc + 1; i++) {
